+++ app/[locale]/components/AddChartButton.tsx
@@ -1,19 +1,8 @@
 import { addChartFunction } from "../actions";
 import { useAppContext } from "../../context/index";
-import { useEffect } from "react";
-import { cartCount } from "../actions";
 
 export default function AddChartButton({ productId }: { productId: number }) {
   const { setState } = useAppContext();
-
-  useEffect(() => {
-    async function getcart() {
-      const count = await cartCount();
-      setState(count);
-    }
-
-    getcart();
-  }, [setState]);
 
   async function handleAddProduct() {
     addChartFunction(productId);
